fix container image scope error:
from add-apt-repository manpage: Adds a repository into the etc.apt.sources.list or etc.apt.sources.list.d or removes an existing one.

But it cant actually edit .etc.apt.sources.list or .etc.apt.sources.list.d , since these two files are in a lower dir in the union mount that overlayfs uses.

Lower dirs are read-only, so even if the command is run as root user... probably it shouldn't work? And that's what throws the error.